backend/requirements.txt
asgiref==3.10.0
cffi==2.0.0
cryptography==46.0.2
Django==5.2.7
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
pycparser==2.23
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
